test: remove assert construct from tests

+++ tests/utilities/test_frontend_builder.py
@@ -34,22 +34,25 @@ class TestGetNodeDirFolder(unittest.TestCase):
     def test_get_installed_node_dir_ok(self):
         node_dir = frontend_builder.get_installed_node()
         app_logger.info(f"node_dir with NODE_DIR:{node_dir}.")
-        assert str(node_dir) == str(node_dir_mocked_bin)
+        if str(node_dir) != str(node_dir_mocked_bin):
+            raise ValueError(f"wrong node folders: get_installed_node '{node_dir}' vs mocked '{node_dir_mocked_bin}' #")
 
     @mock.patch.dict(os.environ, {"NODE_DIR_PARENT": str(node_dir_parent_mocked)})
     def test_get_installed_node_dir_parent_ok(self):
         node_dir = frontend_builder.get_installed_node()
         app_logger.info(f"node_dir with NODE_DIR_PARENT:{node_dir}.")
-        assert str(node_dir) == str(node_dir_mocked_bin)
+        if str(node_dir) != str(node_dir_mocked_bin):
+            raise ValueError(f"wrong node folders: get_installed_node '{node_dir}' vs mocked '{node_dir_mocked_bin}' #")
 
     @mock.patch.dict(os.environ, {"NODE_DIR_PARENT": "", "NODE_DIR": ""})
     def test_get_installed_node_no_env(self):
-        with self.assertRaises(AssertionError):
+        with self.assertRaises(ValueError):
             try:
                 frontend_builder.get_installed_node()
-            except AssertionError as ae:
-                assert "NODE_DIR_PARENT/NODE_DIR env variable not found." in str(ae)
-                raise ae
+            except ValueError as ve:
+                if "NODE_DIR_PARENT/NODE_DIR env variable not found." not in str(ve):
+                    raise AssertionError("ValueError exception hasn't the right message!")
+                raise ve
 
 class TestGetPathWithNodeDir(unittest.TestCase):
     def setUp(self):
@@ -65,18 +68,21 @@ class TestGetPathWithNodeDir(unittest.TestCase):
     @mock.patch.dict(os.environ, {"PATH": "", "NODE_DIR": str(node_dir_mocked)})
     def test_get_path_with_node_dir_path_none(self):
         path = frontend_builder.get_path_with_node_dir()
-        assert f"{node_dir_mocked_bin}:" == path
+        if f"{node_dir_mocked_bin}:" != path:
+            raise ValueError(f"wrong folders: {node_dir_mocked_bin} not in {path} #")
 
     @mock.patch.dict(os.environ, {"PATH": "/bin:", "NODE_DIR": str(node_dir_mocked)})
     def test_get_path_with_node_dir_path_no_node(self):
         path = frontend_builder.get_path_with_node_dir()
-        assert f"{node_dir_mocked_bin}:/bin:" == path
+        if f"{node_dir_mocked_bin}:/bin:" != path:
+            raise ValueError(f"wrong folders: {node_dir_mocked_bin}:/bin: not in {path} #")
 
     @mock.patch.dict(os.environ, {"PATH": f"{node_dir_mocked_bin}:/bin:", "NODE_DIR": str(node_dir_mocked)})
     def test_get_path_with_node_dir_path_no_node__node_dir_mocked_bin_bin(self):
         path = frontend_builder.get_path_with_node_dir()
         app_logger.info(f"PATH '{path}'.")
-        assert f"{node_dir_mocked_bin}:/bin:" == path
+        if f"{node_dir_mocked_bin}:/bin:" != path:
+            raise ValueError(f"wrong folders: {node_dir_mocked_bin}:/bin: not in {path} #")
 
 
 class TestFrontendBuilder(unittest.TestCase):
@@ -84,34 +90,40 @@ class TestFrontendBuilder(unittest.TestCase):
         import shutil
 
         static_dist_folder = static_folder / "dist"
-        assert frontend_builder.build_frontend(
-            project_root_folder=TEST_ROOT_FOLDER,
-            input_css_path=static_css_path,
-            output_dist_folder=static_dist_folder,
-            force_build=True
-        )
-        assert static_dist_folder.is_dir()
+        if not frontend_builder.build_frontend(
+                project_root_folder=TEST_ROOT_FOLDER,
+                input_css_path=static_css_path,
+                output_dist_folder=static_dist_folder,
+                force_build=True
+            ):
+            raise ValueError("Frontend build failed/1")
+        if not static_dist_folder.is_dir():
+            raise FileNotFoundError(f"static_dist_folder folder not found: {static_dist_folder} #")
         index_html = static_dist_folder / "index.html"
         output_css = static_dist_folder / "output.css"
-        assert index_html.is_file()
-        assert output_css.is_file()
+        if not index_html.is_file():
+            raise ValueError(f"missing file: {index_html} #")
+        if not output_css.is_file():
+            raise ValueError(f"missing file: {output_css} #")
         with open(index_html, "r") as html_src:
             html_body = html_src.read()
-            assert "html" in html_body
-            assert "head" in html_body
-            assert "body" in html_body
-        assert output_css.stat().st_size > 0
-        assert not frontend_builder.build_frontend(
-            project_root_folder=TEST_ROOT_FOLDER,
-            input_css_path=static_css_path,
-            output_dist_folder=static_dist_folder
-        )
-        assert not frontend_builder.build_frontend(
-            project_root_folder=TEST_ROOT_FOLDER,
-            input_css_path=static_css_path,
-            output_dist_folder=static_dist_folder,
-            force_build=False
-        )
+            if not ("html" in html_body and "head" in html_body and "body" in html_body):
+                raise ValueError("Wrong frontend build: index html malfolded!")
+        if output_css.stat().st_size < 1:
+            raise IOError("empty output_css file")
+        if frontend_builder.build_frontend(
+                project_root_folder=TEST_ROOT_FOLDER,
+                input_css_path=static_css_path,
+                output_dist_folder=static_dist_folder
+            ):
+            raise ValueError("wrong frontend build/2")
+        if frontend_builder.build_frontend(
+                project_root_folder=TEST_ROOT_FOLDER,
+                input_css_path=static_css_path,
+                output_dist_folder=static_dist_folder,
+                force_build=False
+            ):
+            raise ValueError("wrong frontend build/3")
         shutil.rmtree(static_dist_folder, ignore_errors=False)
